+++ src/nanonext.h
@@ -20,6 +20,7 @@ extern SEXP nano_AioSymbol;
 
 /* define functions */
 extern SEXP rnng_aio_get_msg(SEXP);
+extern SEXP rnng_aio_peek(SEXP);
 extern SEXP rnng_aio_result(SEXP);
 extern SEXP rnng_aio_stop(SEXP);
 extern SEXP rnng_close(SEXP);